fix port 0 giving random port on listener
Basically i was dumb and didn't understand that listening on port 0
made the OS assign you a random port. I thought it made you listen on
all the ports, which in retrospect is dumb because that would be
kind of insane.

Fixed by making a separate TCP listener for all 256 possible TCP
ports the app lets you have at once.